isp-management-system/backend/requirements.txt
# Core Django dependencies
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
redis==5.0.3

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1
cryptography==42.0.5

# RADIUS Integration
pyrad==2.4

# Task Queue
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Utilities
python-dotenv==1.0.1
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2

# Monitoring & Logging
django-debug-toolbar==4.3.0
django-extensions==3.2.3

# API Documentation
drf-yasg==1.21.7

# Payment Processing
stripe==8.2.0

# Network Monitoring
netmiko==4.3.0
paramiko==3.4.0

# Charts and Analytics
django-chartjs==2.3.0

# Export functionality
django-import-export==3.3.7
openpyxl==3.1.2

# Testing
pytest==8.0.2
pytest-django==4.8.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0